requirements.txt
numpy==1.20.3
opencv_python==4.5.5.64
Pillow==9.0.1
pytesseract==0.3.9
